+++ src/typings/cloudDevice/models.ts
@@ -225,8 +225,8 @@ export * from "./storedValueRequest";
 export * from "./storedValueResponse";
 export * from "./storedValueResult";
 export * from "./storedValueTransactionType";
-export * from "./terminalApiRequest";
-export * from "./terminalApiResponse";
+export * from "./cloudDeviceApiRequest";
+export * from "./cloudDeviceApiResponse";
 export * from "./terminalApiSecuredRequest";
 export * from "./terminalApiSecuredResponse";
 export * from "./terminalEnvironmentType";
@@ -459,8 +459,8 @@ import { StoredValueRequest } from "./storedValueRequest";
 import { StoredValueResponse } from "./storedValueResponse";
 import { StoredValueResult } from "./storedValueResult";
 import { StoredValueTransactionType } from "./storedValueTransactionType";
-import { CloudDeviceApiRequest } from "./terminalApiRequest";
-import { CloudDeviceApiResponse } from "./terminalApiResponse";
+import { CloudDeviceApiRequest } from "./cloudDeviceApiRequest";
+import { CloudDeviceApiResponse } from "./cloudDeviceApiResponse";
 import { TerminalApiSecuredRequest } from "./terminalApiSecuredRequest";
 import { TerminalApiSecuredResponse } from "./terminalApiSecuredResponse";
 import { TerminalEnvironmentType } from "./terminalEnvironmentType";
